Moved command package to internal/command
cmd and command are confusing so to disambiguate the two, we now have
our command package moved to `internal/command`.
