--- command/README.md
@@ -1,3 +1,0 @@
-# github.com/cyberark/summon/command
-
-Provides the flags and action for the Summon command-line interface.
