+++ internal/command/README.md
@@ -1,0 +1,3 @@
+# github.com/cyberark/summon/internal/command
+
+Provides the flags and action for the Summon command-line interface.
